Agent1/sdk/pyproject.toml
[project]
name = "sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
]

[project.scripts]
sdk = "sdk:main"
Irfan = "sdk.hello_agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
